Move MenuFrame to sub-module
